[SID:3731] refactor(infra): 공유 i18n 파서를 scripts/ → packages/scripts/ 이관(②) — 근본 봉쇄
story #3731 커밋② — «루트 scripts/를 apps/web이 import하면 Docker 빌드 컨텍스트를
벗어난다» 클래스(story #3729 실사고)를 가드가 «지켜보는» 대신 구조적으로 «못 벌어지게»
막는다. packages/는 Dockerfile builder 스테이지가 이미 COPY하므로(`COPY --from=deps
/app/packages ./packages`), 공유 파서가 거기 있으면 이 클래스 자체가 성립하지 않는다.

이관: scripts/i18n-key-parser.js(+test) → packages/scripts/i18n-key-parser.js(+test).
packages/scripts/는 이미 존재하는 관례(migrate-from-specsite.ts 등, package.json
없이 loose 스크립트가 놓이는 자리) — 새 패키지 신설 0.

소비처 4곳 경로 갱신:
- scripts/check-i18n-keys.js(repo-root, node 직접 실행·로더 없음) — require 경로만
  변경, plain CJS 제약(story #3156 이유) 무변.
- apps/web/scripts/verify-no-new-raw-fetch-api.ts
- apps/web/src/lib/i18n-key-coverage.test.ts
- apps/web/src/lib/pagination-envelope-consumers.test.ts

#3729의 `COPY scripts/ scripts/`(+주석) 걷음 — 이관 뒤에도 남겨 두면 커밋①의 가드
허용 접두에 `scripts/`가 계속 들어가 «루트 scripts/ 재발» 클래스를 그 가드가 영영
못 보게 된다(양성대조 급소). 걷은 뒤 apps/web 비-테스트 전수 재스캔 — 여전히 0건
(이관이 근본 봉쇄임을 실측, 접두 목록에서 `scripts` 자체가 사라짐).

검증 — 가드 재실행(1234파일·0건, scripts 접두 소멸 확認) · vitest 27/27(가드+
verify-no-new-raw-fetch-api+i18n-key-coverage+pagination-envelope-consumers) ·
packages/scripts/i18n-key-parser.test.js 13/13 · tsc --noEmit clean · `node
scripts/check-i18n-keys.js` 정상 실행(exit 0, require 경로 변경 후에도 무회귀).
