+++ packages/scripts/i18n-key-parser.test.js
